gpu: ion: switch to using kmalloc for the secure chunk list
When securing buffers for secure content playback, a list of 1MB
chunks is generated. The number of buffers that will ever be in this
list will never be more than 1024, which means the buffer list will
always be less than a page. Rather than use the contiguous APIs, just
use the slab allocator to get the physical address list. The slab will
be physically contiguous and appropriate to pass to TZ.

Change-Id: Ib3f7293fbe0e5db872c9c441645ca860cc9bb9ff
Signed-off-by: Laura Abbott <lauraa@codeaurora.org>